* Rewrite tests for #474 * Organize junit tree
