--- javaparser-testing/src/test/java/com/github/javaparser/junit/BreadthFirst/D.java
@@ -1,9 +1,0 @@
-package com.github.javaparser.junit.BreadthFirst;
-
-public class D {
-
-    public int foo(int e) {
-        int a = 20;
-        return a;
-    }
-}
