+++ javaparser-testing/src/test/java/com/github/javaparser/junit/ast/visitor/Visitors.java
@@ -1,4 +1,4 @@
-package com.github.javaparser.junit.ast;
+package com.github.javaparser.junit.ast.visitor;
 
 import com.github.javaparser.ast.visitor.GenericVisitorAdapter;
 import com.github.javaparser.ast.visitor.ModifierVisitorAdapter;
